backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
python-dotenv
mysql-connector-python
pydantic[email]
python-multipart
